Rename: ManageContent to RouteButtons
